MP-152 :truck: Rename: normal 디렉토리를 auth 하부로 이동
- normal과 auth가 연관성이 있으므로 디렉토리를 이동함
