+++ src/main/java/kr/modusplant/modules/auth/normal/app/controller/NormalSignUpController.java
@@ -1,4 +1,4 @@
-package kr.modusplant.modules.signup.normal.app.controller;
+package kr.modusplant.modules.auth.normal.app.controller;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.exc.UnrecognizedPropertyException;
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import kr.modusplant.domains.term.domain.model.Term;
 import kr.modusplant.global.app.servlet.response.DataResponse;
-import kr.modusplant.modules.signup.normal.app.http.request.NormalSignUpRequest;
-import kr.modusplant.modules.signup.normal.app.service.NormalSignUpApplicationService;
+import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
+import kr.modusplant.modules.auth.normal.app.service.NormalSignUpApplicationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
