+++ src/main/java/kr/modusplant/modules/auth/normal/mapper/domain/SiteMemberDomainMapper.java
@@ -1,7 +1,7 @@
-package kr.modusplant.modules.signup.normal.mapper.domain;
+package kr.modusplant.modules.auth.normal.mapper.domain;
 
 import kr.modusplant.domains.member.domain.model.SiteMember;
-import kr.modusplant.modules.signup.normal.app.http.request.NormalSignUpRequest;
+import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
 import org.mapstruct.*;
 
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_NULL,
