+++ src/main/java/kr/modusplant/modules/auth/normal/mapper/domain/SiteMemberTermDomainMapper.java
@@ -1,8 +1,8 @@
-package kr.modusplant.modules.signup.normal.mapper.domain;
+package kr.modusplant.modules.auth.normal.mapper.domain;
 
 import kr.modusplant.domains.member.domain.model.SiteMember;
 import kr.modusplant.domains.member.domain.model.SiteMemberTerm;
-import kr.modusplant.modules.signup.normal.app.http.request.NormalSignUpRequest;
+import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.NullValuePropertyMappingStrategy;
